implement proper behavior of 'basedir' config variable during a sequence
If basedir is provided on the sequence level, shell will be told to move to
the basedir in the beginning of the sequence, but won't cd there at the
beginning of each command in the sequence. User will have to override it by
setting basedir explicitly to each command inside the sequence.